requirements.txt
ujson==5.10.0
emoji==2.2.0
rich==13.9.4
unidecode==1.3.6
openai==1.59.8
pyyaml==6.0.2
tenacity==9.0.0
tabulate==0.9.0
pandas==2.2.3
datasets==3.0.0 
tqdm==4.67.1
python-dotenv==1.0.1
openpyxl==3.1.5